coroutine: add a way of propagating exceptions without throwing
It's unfortunate that propagating an exception from a coroutine
can only be done by rethrowing it, if we already hold the proper
exception_ptr instance. In order to make it possible to mitigate
throwing, an explicit mechanism is added. It only works for coroutines
which return future<T> as opposed to future<> due to the fact that
it's illegal to specify both return_void and return_value inside
coroutine traits, but it's better than nothing.
This commit comes with a simple test which also acts as example usage.