websocket: add skeleton implementation of a websocket server
This commit adds minimal foundations for a WebSocket server
implementation for Seastar.
This minimal implementation does not implement parsing
actual WebSocket frames nor sending WebSocket responses,
but it *does* correctly support the handshake protocol,
based on HTTP.
The implementation was tested with Python's websockets package,
which was successfully used to establish a connection
and send arbitrary data to the server.

This commit serves as a kickstarter for the "WebSocket for Seastar"
student project.